+++ src/main/java/com/plivo/api/PlivoAPIService.java
@@ -36,6 +36,7 @@ import com.plivo.api.models.brand.Brand;
 import com.plivo.api.models.brand.BrandCreateResponse;
 import com.plivo.api.models.brand.BrandCreator;
 import com.plivo.api.models.brand.BrandUsecase;
+import com.plivo.api.models.brand.BrandDeleteResponse;
 import com.plivo.api.models.campaign.*;
 import com.plivo.api.models.profile.*;
 import com.plivo.api.models.token.TokenCreateResponse;
@@ -256,12 +257,19 @@ public interface PlivoAPIService {
   @GET("Account/{authId}/10dlc/Brand/{id}/")
   Call<Brand> brandGet(@Path("authId") String authId, @Path("id") String brandId);
 
+  @DELETE("Account/{authId}/10dlc/Brand/{id}/")
+  Call<BrandDeleteResponse> brandDelete(@Path("authId") String authId, @Path("id") String brandId);
+
   @GET("Account/{authId}/10dlc/Brand/{id}/usecases/")
   Call<BrandUsecase> brandUsecaseGet(@Path("authId") String authId, @Path("id") String brandId);
 
   @POST("Account/{authId}/10dlc/Campaign/")
   Call<CampaignCreateResponse> createCampaign(@Path("authId") String authId,
                                           @Body CampaignCreator campaignCreator);
+
+  @POST("Account/{authId}/10dlc/Campaign/{id}/")
+  Call<CampaignUpdateResponse> updateCampaign(@Path("authId") String authId, @Path("id") String campaignId,
+                                          @Body CampaignUpdater campaignUpdater);
 
   @POST("Account/{authId}/10dlc/Campaign/{campaign_id}/Number/")
   Call<CampaignNumberLinkerResponse> linkCampaignNumber(@Path("authId") String authId, @Path("campaign_id") String campaignID,
@@ -304,6 +312,9 @@ public interface PlivoAPIService {
 
   @GET("Account/{authId}/10dlc/Campaign/{campaignId}/")
   Call<Campaign> campaignGet(@Path("authId") String authId, @Path("campaignId") String campaignId);
+
+  @DELETE("Account/{authId}/10dlc/Campaign/{campaignId}/")
+  Call<CampaignDeleteResponse> campaignDelete(@Path("authId") String authId, @Path("campaignId") String campaignId);
 
   // Recording